requirements.txt
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cvxpy==1.1.13
ecos==2.0.7.post1
ez-setup==0.9
Flask==2.0.1
ibm-cloud-networking-services==0.6.0
ibm-cloud-sdk-core==3.10.0
ibm-code-engine-sdk==0.1.0
idna==2.10
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
-e git+https://github.ibm.com/GCAT/cis-integration.git@6a0ee62cbfdb3ba78b43a94cbe8866a92963d961#egg=main.py&subdirectory=src
mapping==0.1.6
MarkupSafe==2.0.1
numpy==1.20.3
osqp==0.6.2.post0
packaging==20.9
pandas==1.2.4
pluggy==0.13.1
py==1.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
qdldl==0.1.5.post0
requests==2.25.1
scipy==1.6.3
scs==2.1.4
six==1.15.0
toml==0.10.2
urllib3==1.26.5
Werkzeug==2.0.1
